requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit_surprise==1.1.3
streamlit==1.23.1
openpyxl==3.0.7